+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.2.0] - Release
+
+* `PhoneNumber` now has a `Country` attribute instead of a simple country code. Allows to access the flag and name of the country afterwards.
+* Added support of a target length, by default `10`.
+* Added support for a counter
+* Added support for custom formatters
+  
+
 ## [1.1.0] - Release
 
 * Added styling of the country selection button.